Add a cache for "config"
Because getAvailableCommands() is used as the basis for many other
actions, and the payload is so gigantic, plus the fact that it does
not change very frequently, it makes sense to keep that response
cached.

The only way this would become out of sync is if you make changes to
the harmony setup (activities and devices).